requirements.txt
# Core web framework
Django>=6.0.0
djangorestframework>=3.16.0

# Authentication
djangorestframework-simplejwt>=5.5.0
PyJWT>=2.11.0

# Environment variables
python-dotenv>=1.2.0

# Django runtime dependencies
asgiref>=3.11.0
sqlparse>=0.5.0
typing_extensions>=4.15.0

# Async job pipeline (agentic matching)
celery>=5.4.0
redis>=5.2.0

# AI/LLM integration
openai>=1.0.0

# Timezone data (required for Windows)
tzdata>=2025.0
